Separating CLI from Exports and exporting utility function
